Rearange module files
