+++ client/blocks/plugins-update-manager/index.tsx
@@ -1,144 +1,42 @@
-import {
-	__experimentalText as Text,
-	__experimentalConfirmDialog as ConfirmDialog,
-	Button,
-	Card,
-	CardBody,
-	CardHeader,
-	DropdownMenu,
-	Tooltip,
-} from '@wordpress/components';
-import { createInterpolateElement } from '@wordpress/element';
-import { Icon, arrowLeft, info, plus } from '@wordpress/icons';
-import { useState } from 'react';
-import { ellipsis } from './icons';
-
-import './styles.scss';
+import { Button } from '@wordpress/components';
+import { plus } from '@wordpress/icons';
+import MainComponent from 'calypso/components/main';
+import NavigationHeader from 'calypso/components/navigation-header';
+import { ScheduleCreate } from './schedule-create';
+import { ScheduleList } from './schedule-list';
 
 interface Props {
+	siteSlug: string;
+	context: 'list' | 'create';
 	onNavBack?: () => void;
+	onCreateNewSchedule?: () => void;
 }
 export const PluginsUpdateManager = ( props: Props ) => {
-	const { onNavBack } = props;
-	const [ isConfirmOpen, setIsConfirmOpen ] = useState( false );
-
-	const closeConfirm = () => {
-		setIsConfirmOpen( false );
-	};
-
-	const toolbarPluginsText = createInterpolateElement(
-		'<div>Move to WordPress.com<br />Akismet<br />Gravity Forms</div>',
-		{
-			div: <div className="tooltip--selected-plugins" />,
-			br: <br />,
-		}
-	);
+	const { context, onNavBack, onCreateNewSchedule } = props;
 
 	return (
-		<>
-			<ConfirmDialog isOpen={ isConfirmOpen } onConfirm={ closeConfirm } onCancel={ closeConfirm }>
-				Are you sure you want to delete this schedule?
-			</ConfirmDialog>
-			<Card className="plugins-update-manager">
-				<CardHeader size="extraSmall">
-					<div className="ch-placeholder">
-						{ onNavBack && (
-							<Button icon={ arrowLeft } onClick={ onNavBack }>
-								Back
-							</Button>
-						) }
-					</div>
-					<Text>Schedules</Text>
-					<div className="ch-placeholder"></div>
-				</CardHeader>
-				<CardBody>
-					<div className="empty-state">
-						<Text as="p" align="center">
-							Set up plugin update schedules to ensure your site runs smoothly.
-						</Text>
-						<Button __next40pxDefaultSize icon={ plus } variant="primary">
-							Create a new schedule
-						</Button>
-					</div>
-				</CardBody>
-				<CardBody>
-					<table>
-						<thead>
-							<tr>
-								<th>Name</th>
-								<th>Next Update</th>
-								<th>Frequency</th>
-								<th>Plugins</th>
-								<th></th>
-							</tr>
-						</thead>
-						<tbody>
-							<tr>
-								<td className="name">Move to WordPress.com plugin</td>
-								<td>Feb 28 2023 7:00 PM UTC</td>
-								<td>Daily</td>
-								<td>
-									1
-									<Tooltip
-										text="Move to WordPress.com plugin"
-										position="middle right"
-										delay={ 0 }
-										hideOnClick={ false }
-									>
-										<Icon className="icon-info" icon={ info } size={ 16 } />
-									</Tooltip>
-								</td>
-								<td style={ { textAlign: 'end' } }>
-									<DropdownMenu
-										popoverProps={ { position: 'bottom left' } }
-										controls={ [
-											{
-												title: 'Remove',
-												onClick: () => setIsConfirmOpen( true ),
-											},
-										] }
-										icon={ ellipsis }
-										label="More"
-									/>
-								</td>
-							</tr>
-							<tr>
-								<td className="name">Security plugins</td>
-								<td>Feb 28 2023 7:00 PM UTC</td>
-								<td>Daily</td>
-								<td>
-									3
-									<Tooltip
-										text={ toolbarPluginsText as unknown as string }
-										position="middle right"
-										delay={ 0 }
-										hideOnClick={ false }
-									>
-										<Icon className="icon-info" icon={ info } size={ 16 } />
-									</Tooltip>
-								</td>
-								<td style={ { textAlign: 'end' } }>
-									<DropdownMenu
-										popoverProps={ { position: 'bottom left' } }
-										controls={ [
-											{
-												title: 'Remove',
-												onClick: () => setIsConfirmOpen( true ),
-											},
-										] }
-										icon={ ellipsis }
-										label="More"
-									/>
-								</td>
-							</tr>
-						</tbody>
-					</table>
-					<Text as="p">
-						<Icon className="icon-info" icon={ info } size={ 16 } />
-						The current feature implementation only allows to set up two schedules.
-					</Text>
-				</CardBody>
-			</Card>
-		</>
+		<MainComponent wideLayout>
+			<NavigationHeader
+				navigationItems={ [] }
+				title="Plugin updates manager"
+				subtitle="Effortlessly schedule plugin auto-updates with built-in rollback logic."
+			>
+				{ context === 'list' && onCreateNewSchedule && (
+					<Button
+						__next40pxDefaultSize
+						icon={ plus }
+						variant="primary"
+						onClick={ onCreateNewSchedule }
+					>
+						Create a new schedule
+					</Button>
+				) }
+			</NavigationHeader>
+
+			{ context === 'list' && (
+				<ScheduleList onNavBack={ onNavBack } onCreateNewSchedule={ onCreateNewSchedule } />
+			) }
+			{ context === 'create' && <ScheduleCreate onNavBack={ onNavBack } /> }
+		</MainComponent>
 	);
 };
