+++ client/my-sites/plugins/controller.js
@@ -1,11 +1,11 @@
 import page from '@automattic/calypso-router';
 import { includes, some } from 'lodash';
 import { createElement } from 'react';
+import { PluginsUpdateManager } from 'calypso/blocks/plugins-update-manager';
 import { redirectLoggedOut } from 'calypso/controller';
 import { gaRecordEvent } from 'calypso/lib/analytics/ga';
 import { getSiteFragment, sectionify } from 'calypso/lib/route';
 import { navigation } from 'calypso/my-sites/controller';
-import { UpdatesManager, UpdatesManagerCreate } from 'calypso/my-sites/plugins/updates-manager';
 import { isUserLoggedIn, getCurrentUserSiteCount } from 'calypso/state/current-user/selectors';
 import getSelectedOrAllSitesWithPlugins from 'calypso/state/selectors/get-selected-or-all-sites-with-plugins';
 import { fetchSitePlans } from 'calypso/state/sites/plans/actions';
@@ -117,16 +117,18 @@ export function updatesManager( context, next ) {
 
 	switch ( context.params.action ) {
 		case 'create':
-			context.primary = createElement( UpdatesManagerCreate, {
+			context.primary = createElement( PluginsUpdateManager, {
 				siteSlug,
+				context: 'create',
 				onNavBack: () => page.redirect( `/plugins/update-manager/${ siteSlug }` ),
 			} );
 			break;
 
 		case 'list':
 		default:
-			context.primary = createElement( UpdatesManager, {
+			context.primary = createElement( PluginsUpdateManager, {
 				siteSlug,
+				context: 'list',
 				onCreateNewSchedule: () => page.redirect( `/plugins/update-manager/create/${ siteSlug }` ),
 			} );
 			break;
